Added detected doctype to the error report
It is good to know what standard was used to check the rendered document,
previously that had to be guessed from the doctype in the document itself
after an error was reported. By adding the detected doctype to the report
it is easier to find errors in rendering (e.g. wrong doctype)