Fixes #5 and #8: validator runs for all pages and is present when valid
Wicket doesn't run the response filters when rendering the buffered
response but when sending the buffered response to the browser. This
means that the validator doesn't know which page was rendered, and that
caused the validator to not run in those instances. Unfortunately this
meant that the validator became pretty much useless, as most links and
buttons render to a buffered response.

When a page has been validated and the markup is valid (X)HTML(5), the
validator should be visible, but not obtrusive. This commit adds a
validator button at the bottom of the screen and when clicked, the
validator window becomes visible.